% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSD758-package.R
\docType{package}
\name{OSD758-package}
\alias{OSD758}
\alias{OSD758-package}
\title{OSD758: Artificial Gravity Attenuates the Transcriptomic Response to Spaceflight in the Optic Nerve and Retina}
\description{
Provides sample metadata, raw counts, and normalized transcriptomic data from a NASA-funded experiment examining the effects of artificial gravity on ocular tissues during spaceflight. The dataset originates from the Open Science Data Repository study OSD-758, which includes RNA-seq profiles of mouse retina following exposure to varying gravity levels aboard the International Space Station. This package enables streamlined access and analysis of the data for reproducible research and teaching in space biology, transcriptomics, and gravitational physiology. Source: \url{https://doi.org/10.26030/d6dj-d777}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.github.com/patterninstitute/OSD758/}
  \item \url{https://www.pattern.institute/OSD758/}
  \item Report bugs at \url{https://www.github.com/patterninstitute/OSD758/issues}
}

}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Authors:
\itemize{
  \item Isabel Duarte \email{iduarte@pattern.institute} (\href{https://orcid.org/0000-0003-0060-2936}{ORCID})
}

Other contributors:
\itemize{
  \item Pattern Institute (04jrgd746) [copyright holder, funder]
}

}
\keyword{internal}
